chore: reorganize lsp code to be laid out more intentionally
