+++ tooling/lsp/Cargo.toml
@@ -29,4 +29,4 @@ async-lsp = { version = "0.0.5", default-features = false, features = ["omni-tra
 wasm-bindgen.workspace = true
 
 [dev-dependencies]
-tokio = { version = "1.0", features = ["macros"] }
+tokio = { version = "1.0", features = ["macros", "rt"] }
